Fixed a bug
Try to assign a negative value in the last version and get failed tests.

Bug was: Couldn't create AOS and assign a negative value to all of them directly. This was because of a typo in line 52, where we used "size_t" instead of "arrType". Since "size_t" doesn't allow negative values, we consider this a bug. We are very sorry, and the bug is fixed now.